make nanval more unique

+++ skyportal/handlers/api/photometry.py
@@ -135,7 +135,7 @@ def hash_photometry(row):
     if row['standardized_flux'] is not None:
         m.update(np.float32(row["standardized_flux"]))
     else:
-        m.update(np.float32(np.nan))
+        m.update(np.float128(np.nan))
     m.update(np.float32(row["standardized_fluxerr"]))
     if row["origin"] is not None:
         m.update(row["origin"].encode('utf-8'))
